Created a separate file for deletePhoto. #103
